requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
sdv
plotly
scikit-learn
scipy
fpdf
python-dotenv
pydantic
openpyxl
sqlalchemy
PyPDF2
sentence-transformers
faiss-cpu
anthropic
python-jose[cryptography]
passlib[bcrypt]
argon2-cffi
pyarrow
fhir.resources